Move wasm logic to its own crate
