+++ Cargo.toml
@@ -47,8 +47,10 @@ members = [
   "oak_functions/sdk/oak_functions/tests/metrics_module",
   "oak_functions/sdk/oak_functions/tests/tf_module",
   "oak_functions/sdk/test_utils",
+  "oak_functions/testing",
   "oak_functions/util",
   "oak_functions/workload_logging",
+  "oak_functions/wasm",
   "oak_idl",
   "remote_attestation/rust",
   "remote_attestation_sessions",
