+++ oak_functions/loader/Cargo.toml
@@ -43,6 +43,7 @@ oak_functions_lookup = { path = "../lookup" }
 oak_functions_util = { path = "../util" }
 oak_functions_metrics = { optional = true, path = "../experimental/metrics" }
 oak_functions_tf_inference = { optional = true, path = "../experimental/tf_inference" }
+oak_functions_wasm = { path = "../wasm" }
 oak_functions_workload_logging = { path = "../workload_logging" }
 oak_logger = { path = "../logger" }
 oak_utils = { path = "../../oak_utils" }
