+++ oak_functions/testing/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "oak_functions_testing_extension"
+version = "0.1.0"
+authors = ["Conrad Grobler <grobler@google.com>"]
+edition = "2021"
+license = "Apache-2.0"
+
+[dependencies]
+anyhow = "*"
+bincode = "*"
+chrono = "*"
+oak_logger = { path = "../logger" }
+oak_functions_abi = { path = "../abi" }
+oak_functions_extension = { path = "../extension" }
